Rename and move Loony Tune into the Road Runner folder.
